Project/LLM/requirements.txt
torch
torchvision
torchaudio
tensorflow
flax
gradio
llama-index-embeddings-mixedbreadai
llama-index-llms-groq
llama-parse